me quedo pillado

+++ ProyectoAula2/app/src/main/java/com/example/proyectoaula2/main2/CursosAula.kt
@@ -6,5 +6,5 @@ data class CursosAula(
     val direccion: String,
     val id: Int,
     val nombre: String,
-    val precio: Double
+    val precio: Float
 )
