+++ ProyectoAula2/app/src/main/java/com/example/proyectoaula2/main2/CursosResponse.kt
@@ -6,5 +6,5 @@ data class CursosResponse(
     val direccion: String,
     val id: Int,
     val nombre: String,
-    val precio: Double
+    val precio: Float
 )
